use uintptr_t as value_t

+++ src/types.h
@@ -1,6 +1,6 @@
 typedef uint8_t byte_t;
 typedef byte_t instruction_t;
-typedef uint64_t value_t;
+typedef uintptr_t value_t;
 typedef int64_t int_t;
 
 typedef struct _dict_t dict_t;
